dashboard/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pyyaml==6.0.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.10.3
